+++ shortcodes/stable_diffusion/img2img.py
@@ -17,18 +17,21 @@ class Shortcode():
 				sampler_index = i
 				break
 
+		init_img = self.Unprompted.shortcode_user_vars["init_images"][len(self.Unprompted.after_processed.images) - 1]
+		init_img_with_mask = self.Unprompted.shortcode_user_vars["init_img_with_mask"] if "init_img_with_mask" in self.Unprompted.shortcode_user_vars else None
+
 		img2img_result = modules.img2img.img2img(
-			"unprompted_img2img",
-			self.Unprompted.shortcode_user_vars["img2img_mode"] if "img2img_mode" in self.Unprompted.shortcode_user_vars else 0, #p.mode
+			"unprompted_img2img", #id_task
+			self.Unprompted.shortcode_user_vars["mode"] if "mode" in self.Unprompted.shortcode_user_vars else 0, #p.mode
 			self.Unprompted.shortcode_user_vars["prompt"],
 			self.Unprompted.shortcode_user_vars["negative_prompt"],
-			"None", # prompt_styles
-			self.Unprompted.shortcode_user_vars["init_images"][len(self.Unprompted.after_processed.images) - 1],
+			[], # prompt_styles
+			init_img,
 			None, # sketch
-			None, # p.init_img_with_mask
+			init_img_with_mask, # p.init_img_with_mask
 			None, # inpaint_color_sketch
 			None, # inpaint_color_sketch_orig
-			self.Unprompted.shortcode_user_vars["init_images"][len(self.Unprompted.after_processed.images) - 1], # p.init_img_inpaint
+			init_img, # p.init_img_inpaint
 			init_mask, # p.init_mask_inpaint
 			self.Unprompted.shortcode_user_vars["steps"],
 			sampler_index,
@@ -37,10 +40,11 @@ class Shortcode():
 			0, # p.inpainting_fill
 			self.Unprompted.shortcode_user_vars["restore_faces"],
 			self.Unprompted.shortcode_user_vars["tiling"],
-			self.Unprompted.shortcode_user_vars["n_iter"] if "n_iter" in self.Unprompted.shortcode_user_vars else 1, #p.n_iter
+			self.Unprompted.shortcode_user_vars["n_iter"] if "n_iter" in self.Unprompted.shortcode_user_vars else 1, #p.n_iter - batch count
 			self.Unprompted.shortcode_user_vars["batch_size"] if "batch_size" in self.Unprompted.shortcode_user_vars else 1, #p.batch_size
 			self.Unprompted.shortcode_user_vars["cfg_scale"],
-			self.Unprompted.shortcode_user_vars["denoising_strength"] if self.Unprompted.shortcode_user_vars["denoising_strength"] is not None else 0.5,
+			self.Unprompted.shortcode_user_vars["image_cfg_scale"] if "image_cfg_scale" in self.Unprompted.shortcode_user_vars else None,
+			self.Unprompted.shortcode_user_vars["denoising_strength"] if self.Unprompted.shortcode_user_vars["denoising_strength"] is not None else 1.0,
 			self.Unprompted.shortcode_user_vars["seed"],
 			self.Unprompted.shortcode_user_vars["subseed"],
 			self.Unprompted.shortcode_user_vars["subseed_strength"],
@@ -49,16 +53,19 @@ class Shortcode():
 			0, # seed_enable_extras
 			self.Unprompted.shortcode_user_vars["height"],
 			self.Unprompted.shortcode_user_vars["width"],
-			self.Unprompted.shortcode_user_vars["resize_mode"] if "resize_mode" in self.Unprompted.shortcode_user_vars else 0,
+			self.Unprompted.shortcode_user_vars["resize_mode"] if "resize_mode" in self.Unprompted.shortcode_user_vars else 1,
 			self.Unprompted.shortcode_user_vars["inpaint_full_res"] if "inpaint_full_res" in self.Unprompted.shortcode_user_vars else True, # p.inpaint_full_res
-			self.Unprompted.shortcode_user_vars["inpaint_full_res_padding"] if "inpaint_full_res_pdding" in self.Unprompted.shortcode_user_vars else 0, # p.inpaint_full_res_padding
+			self.Unprompted.shortcode_user_vars["inpaint_full_res_padding"] if "inpaint_full_res_pdding" in self.Unprompted.shortcode_user_vars else 1, # p.inpaint_full_res_padding
 			0, # p.inpainting_mask_invert
 			"", #p.batch_input_directory
 			"", #p.batch_output_directory
 			"", #p.img2img_batch_inpaint_mask_dir
-			0, # this is the *args tuple and I believe 0 indicates we are not using an extra script in img2img
+			"", # override_settings_texts
+			0, # this is the *args tuple, 0 prevents additional scripts from running here
+			0, # Prevents endless loop
+			-1
 		)
-
+		
 		# Get the image stored in the first index
 		img2img_images = img2img_result[0]